# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.CreateChapterPayload:
    properties:
      chapter_number:
        type: number
      content:
        type: string
      is_locked:
        type: boolean
      price:
        type: integer
      slug:
        maxLength: 100
        type: string
      title:
        type: string
    required:
    - chapter_number
    - content
    - slug
    - title
    type: object
...
  main.UpdateChapterPayload:
    properties:
      chapter_number:
        type: number
      content:
        type: string
      is_locked:
        type: boolean
      price:
        type: integer
      title:
        type: string
    type: object
...
  store.Chapter:
    properties:
      chapter_number:
        type: number
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_locked:
        type: boolean
      is_read:
        type: boolean
      novel_id:
        type: integer
      price:
        type: integer
      slug:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  store.Invoice:
    properties:
      amount:
        type: number
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: integer
      invoice_id:
        type: string
      invoice_url:
        type: string
      plan:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
...
  store.UserUnlock:
    properties:
      chapter_slug:
        type: string
      created_at:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  swagger.EnvelopeError:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for an online novel platform — read, publish, and manage
    digital stories.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Govel API
paths:
...
  /invoices:
    get:
      description: Get user's invoices
      produces:
      - application/json
      responses:
        "200":
          description: Detail chapter
          schema:
            $ref: '#/definitions/store.Invoice'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Invoice not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Get invoices
      tags:
      - invoices
  /invoices/{plan}:
    post:
      description: Get invoices according from plan
      parameters:
      - description: Plan
        in: path
        name: plan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detail chapter
          schema:
            $ref: '#/definitions/store.Invoice'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Create invoice
      tags:
      - invoices
  /invoices/all:
    get:
      description: Get all invoices. Admin only
      produces:
      - application/json
      responses:
        "200":
          description: Detail chapter
          schema:
            $ref: '#/definitions/store.Invoice'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Get all invoices
      tags:
      - invoices
...
  /novels/{novelID}/chapters:
    post:
      consumes:
      - application/json
      description: Create a new chapter with slug, title, author, content, chapter
        number, price and status is locked. Admin only
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      - description: chapter payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateChapterPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Chapter created successfully
          schema:
            $ref: '#/definitions/store.Chapter'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Create a new chapter
      tags:
      - novels
  /novels/{novelID}/chapters/{slug}:
    delete:
      consumes:
      - application/json
      description: Delete chapter by slug
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      - description: Chapter Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Delete chapter succesfully
          schema:
            type: ""
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel or chapter not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Delete chapter
      tags:
      - novels
    get:
      description: Get detailed information about a specific chapter by its slug
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      - description: Chapter Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detail chapter
          schema:
            $ref: '#/definitions/store.Chapter'
        "400":
          description: Invalid novel ID or slug
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "402":
          description: Payment required
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel or chapter not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Get chapter detail
      tags:
      - novels
    patch:
      consumes:
      - application/json
      description: Update an existing chapter's title, content, chapter number, status
        is_locked or price. Admin only.
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      - description: Chapter Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateChapterPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Updated chapter
          schema:
            $ref: '#/definitions/store.Chapter'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Update chapter
      tags:
      - novels
  /novels/{novelID}/chapters/{slug}/unlock:
    post:
      description: User can unlock chapter by coin
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      - description: Chapter Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated chapter
          schema:
            $ref: '#/definitions/store.UserUnlock'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "402":
          description: Insufficient coin
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel or chapter not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Unlock chapter
      tags:
      - novels
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"